+++ react/Product.js
@@ -170,9 +170,10 @@ export const parseToJsonLD = ({
   decimals,
   pricesWithTax,
   useSellerDefault,
+  useImagesArray,
   disableAggregateOffer,
 }) => {
-  const [image] = selectedItem ? selectedItem.images : []
+  const images = selectedItem ? selectedItem.images : []
   const { brand } = product
   const name = product.productName
 
@@ -204,7 +205,9 @@ export const parseToJsonLD = ({
     '@id': `${baseUrl}/${product.linkText}/p`,
     name,
     brand: parseBrand(brand),
-    image: image?.imageUrl || null,
+    image: useImagesArray
+      ? images.map((el) => el.imageUrl)
+      : images[0]?.imageUrl || null,
     description: product.metaTagDescription || product.description,
     mpn,
     sku: selectedItem?.itemId || null,
@@ -226,6 +229,7 @@ function StructuredData({ product, selectedItem }) {
     disableOffers,
     pricesWithTax,
     useSellerDefault,
+    useImagesArray,
     disableAggregateOffer,
   } = useAppSettings()
 
@@ -237,6 +241,7 @@ function StructuredData({ product, selectedItem }) {
     decimals,
     pricesWithTax,
     useSellerDefault,
+    useImagesArray,
     disableAggregateOffer,
   })